requirements.txt
langchain
streamlit
openai
PyPDF
faiss-cpu
tiktoken
python-dotenv
google-generativeai